Include ColumnParameter as a nested class in the Base data set
